refactor :: group list 개선
