+++ src/main/kotlin/team/kin/forest/domain/group/adapter/input/mapper/GroupDataMapper.kt
@@ -1,15 +1,23 @@
 package team.kin.forest.domain.group.adapter.input.mapper
 
 import org.springframework.stereotype.Component
-import team.kin.forest.domain.group.adapter.input.data.response.QueryGroupResponse
+import team.kin.forest.domain.group.adapter.input.data.response.QueryGroupsResponse
 import team.kin.forest.domain.group.application.port.output.dto.GroupsDto
+import team.kin.forest.domain.group.adapter.input.data.response.QueryGroupsResponse.QueryGroupResponse as QueryGroupResponse
 
 @Component
 class GroupDataMapper {
-    infix fun toResponse(dto: GroupsDto) = QueryGroupResponse(
-        id = dto.id,
-        name = dto.name,
-        headcount = dto.headcount,
-        todo = dto.todo
-    )
+
+    infix fun toResponse(dtos: List<GroupsDto>) =
+        QueryGroupsResponse(
+            dtos.map {
+                QueryGroupResponse(
+                    id = it.id,
+                    name = it.name,
+                    headcount = it.headcount,
+                    todo = it.todo
+                )
+            }
+        )
+
 }
